:recycle: Split wasm::layout into submodules
