+++ render-wasm/src/wasm.rs
@@ -1,7 +1,7 @@
 pub mod blurs;
 pub mod fills;
 pub mod fonts;
-pub mod layout;
+pub mod layouts;
 pub mod paths;
 pub mod shapes;
 pub mod strokes;
