+++ render-wasm/src/wasm/layouts.rs
@@ -1,121 +1,9 @@
 use crate::mem;
-use crate::shapes::{self, ConstraintH, ConstraintV, VerticalAlign};
+use crate::shapes::{self};
 use crate::{uuid_from_u32_quartet, with_current_shape_mut, with_state, with_state_mut, STATE};
-use macros::ToJs;
 
-#[derive(Debug, Clone, PartialEq, Copy, ToJs)]
-#[repr(u8)]
-#[allow(dead_code)]
-pub enum RawConstraintH {
-    Left = 0,
-    Right = 1,
-    Leftright = 2, // odd casing to comply with cljs value
-    Center = 3,
-    Scale = 4,
-}
-
-impl From<u8> for RawConstraintH {
-    fn from(value: u8) -> Self {
-        unsafe { std::mem::transmute(value) }
-    }
-}
-
-impl From<RawConstraintH> for ConstraintH {
-    fn from(value: RawConstraintH) -> Self {
-        match value {
-            RawConstraintH::Left => ConstraintH::Left,
-            RawConstraintH::Right => ConstraintH::Right,
-            RawConstraintH::Leftright => ConstraintH::LeftRight,
-            RawConstraintH::Center => ConstraintH::Center,
-            RawConstraintH::Scale => ConstraintH::Scale,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Copy, ToJs)]
-#[repr(u8)]
-#[allow(dead_code)]
-pub enum RawConstraintV {
-    Top = 0,
-    Bottom = 1,
-    Topbottom = 2, // odd casing to comply with cljs value
-    Center = 3,
-    Scale = 4,
-}
-
-impl From<u8> for RawConstraintV {
-    fn from(value: u8) -> Self {
-        unsafe { std::mem::transmute(value) }
-    }
-}
-
-impl From<RawConstraintV> for ConstraintV {
-    fn from(value: RawConstraintV) -> Self {
-        match value {
-            RawConstraintV::Top => ConstraintV::Top,
-            RawConstraintV::Bottom => ConstraintV::Bottom,
-            RawConstraintV::Topbottom => ConstraintV::TopBottom,
-            RawConstraintV::Center => ConstraintV::Center,
-            RawConstraintV::Scale => ConstraintV::Scale,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Copy, ToJs)]
-#[repr(u8)]
-#[allow(dead_code)]
-pub enum RawVerticalAlign {
-    Top = 0,
-    Center = 1,
-    Bottom = 2,
-}
-
-impl From<u8> for RawVerticalAlign {
-    fn from(value: u8) -> Self {
-        unsafe { std::mem::transmute(value) }
-    }
-}
-
-impl From<RawVerticalAlign> for VerticalAlign {
-    fn from(value: RawVerticalAlign) -> Self {
-        match value {
-            RawVerticalAlign::Top => VerticalAlign::Top,
-            RawVerticalAlign::Center => VerticalAlign::Center,
-            RawVerticalAlign::Bottom => VerticalAlign::Bottom,
-        }
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn set_shape_constraint_h(constraint: u8) {
-    with_current_shape_mut!(state, |shape: &mut Shape| {
-        let constraint = RawConstraintH::from(constraint);
-        shape.set_constraint_h(Some(constraint.into()));
-    });
-}
-
-#[no_mangle]
-pub extern "C" fn set_shape_constraint_v(constraint: u8) {
-    with_current_shape_mut!(state, |shape: &mut Shape| {
-        let constraint = RawConstraintV::from(constraint);
-        shape.set_constraint_v(Some(constraint.into()));
-    });
-}
-
-#[no_mangle]
-pub extern "C" fn clear_shape_constraints() {
-    with_current_shape_mut!(state, |shape: &mut Shape| {
-        shape.clear_constraints();
-    });
-}
-
-#[no_mangle]
-pub extern "C" fn set_shape_vertical_align(align: u8) {
-    with_current_shape_mut!(state, |shape: &mut Shape| {
-        let align = RawVerticalAlign::from(align);
-        shape.set_vertical_align(align.into());
-    });
-}
+mod align;
+mod constraints;
 
 #[no_mangle]
 pub extern "C" fn clear_shape_layout() {
